requirements.txt
streamlit
httpx
base64
dotenv
google.generativeai